+++ packages/storybook/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @stcland/storybook
 
+## 0.0.29
+
+### Patch Changes
+
+- Added expand/collapse all in Accordions.
+- Updated dependencies
+  - @stcland/components@0.0.25
+
 ## 0.0.28
 
 ### Patch Changes